+++ lib/ice/output/ice_puts.c
@@ -6,9 +6,10 @@
 */
 
 #include <unistd.h>
+
 #include "ice/string.h"
 
-int ice_puts(const char *str)
+ssize_t ice_puts(const char *str)
 {
     return write(1, str, ice_strlen(str));
 }
