+++ lib/ice/string/ice_strcpy.c
@@ -5,9 +5,11 @@
 ** ice_strcpy.c
 */
 
+#include "ice/types.h"
+
 void ice_strcpy(char *dest, const char *src)
 {
-    int i = 0;
+    ULL i = 0;
 
     for (; src[i] ; i++)
         dest[i] = src[i];
