+++ lib/list/list_remove.c
@@ -6,9 +6,10 @@
 */
 
 #include <stdlib.h>
+
 #include "list.h"
 
-void *list_remove(list_t *list, unsigned int index)
+void *list_remove(list_t *list, ULL index)
 {
     void *value;
     list_node_t *node;
@@ -17,7 +18,7 @@ void *list_remove(list_t *list, unsigned int index)
         return NULL;
 
     node = list->head;
-    for (unsigned int i = 0; i < index; i++)
+    for (ULL i = 0; i < index; i++)
         node = node->next;
 
     list_pop_node(list, node);
